+++ src/Client/ClientCore/Systems/WorldManagement/ClientWorldSegmentLoader.cs
@@ -26,10 +26,12 @@ using MessagePack;
 using Sovereign.ClientCore.Network.Rest;
 using Sovereign.EngineCore.Components.Indexers;
 using Sovereign.EngineCore.Events;
+using Sovereign.EngineCore.Network;
 using Sovereign.EngineCore.Network.Rest;
 using Sovereign.EngineCore.Systems.Block;
 using Sovereign.EngineCore.Systems.Block.Events;
 using Sovereign.EngineCore.Systems.WorldManagement;
+using Sovereign.EngineCore.World.Materials;
 using Sovereign.WorldManagement.Configuration;
 using Sovereign.WorldManagement.Systems.WorldManagement;
 using Sovereign.WorldManagement.WorldSegments;
@@ -83,11 +85,6 @@ namespace Sovereign.ClientCore.Systems.WorldManagement
 
         public void LoadSegment(GridPosition segmentIndex)
         {
-            /* Fill the segment with a known block pattern for test. */
-            blockController.AddBlocks(eventSender,
-                (blocks) => CreateBlocks(blocks, segmentIndex));
-            worldSegmentRegistry.OnSegmentLoaded(segmentIndex);
-
             /* Spin up a background task to asynchronously get the segment from the server. */
             Task.Run(() => RetrieveSegmentViaRest(segmentIndex));
         }
@@ -123,44 +120,44 @@ namespace Sovereign.ClientCore.Systems.WorldManagement
                     {
                         case HttpStatusCode.OK:
                             var segmentBytes = await response.Content.ReadAsByteArrayAsync();
-                            ProcessSegmentBytes(segmentBytes);
+                            ProcessSegmentBytes(segmentIndex, segmentBytes);
                             worldSegmentRegistry.OnSegmentLoaded(segmentIndex);
                             return;
 
                         case HttpStatusCode.ServiceUnavailable:
                             // Server has not yet loaded the segment - delay and try again.
-                            Logger.WarnFormat("Server is not ready to serve segment data for {1}.", segmentIndex);
+                            Logger.WarnFormat("Server is not ready to serve segment data for {0}.", segmentIndex);
                             break;
 
                         default:
                             // Other unknown error - log, try again.
-                            Logger.ErrorFormat("Unexpected response code {1} retrieving segment data for {2}.", response.StatusCode, segmentIndex);
+                            Logger.ErrorFormat("Unexpected response code {0} retrieving segment data for {1}.", response.StatusCode, segmentIndex);
                             break;
                     }
                 }
                 catch (Exception e)
                 {
-                    Logger.ErrorFormat(e, "Exception while retrieving segment data for {1}.", segmentIndex);
+                    Logger.ErrorFormat(e, "Exception while retrieving segment data for {0}.", segmentIndex);
                 }
 
                 retryCount++;
             }
 
             /* If we get here, all the retries failed and we hit the limit. */
-            Logger.ErrorFormat("Failed to retrieve segment data for {1} after {2} tries; connection assumed lost.", segmentIndex, MaxRetries);
+            Logger.ErrorFormat("Failed to retrieve segment data for {0} after {1} tries; connection assumed lost.", segmentIndex, MaxRetries);
             // TODO: Send an event indicating presumed connection failure.
         }
 
         /// <summary>
         /// Loads a segment from its compressed representation.
         /// </summary>
+        /// <param name="segmentIndex">Segment index.</param>
         /// <param name="segmentBytes">Bytes from server.</param>
-        /// <exception cref="NotImplementedException"></exception>
-        private void ProcessSegmentBytes(byte[] segmentBytes)
+        private void ProcessSegmentBytes(GridPosition segmentIndex, byte[] segmentBytes)
         {
             /* Decompress and unpack data. */
-            var options = MessagePackSerializerOptions.Standard.WithCompression(MessagePackCompression.Lz4BlockArray);
-            var segmentData = MessagePackSerializer.Deserialize<WorldSegmentBlockData>(segmentBytes, options);
+            var segmentData = MessagePackSerializer.Deserialize<WorldSegmentBlockData>(segmentBytes, 
+                MessageConfig.CompressedUntrustedMessagePackOptions);
 
             /* Basic top-level validation. */
             if (segmentData.DefaultMaterialsPerPlane.Length != config.SegmentLength)
@@ -168,8 +165,23 @@ namespace Sovereign.ClientCore.Systems.WorldManagement
                 throw new RankException("Bad number of default materials in segment data.");
             }
 
+            /* Submit all the blocks we've added. */
+            var controller = new BlockController();
+            controller.AddBlocks(eventSender, (blocksToAdd) => CreateBlocksToAdd(segmentIndex, segmentData, blocksToAdd));
+        }
+
+        /// <summary>
+        /// Creates block records from segment data and adds them to the given list.
+        /// </summary>
+        /// <param name="segmentIndex">Segment index.</param>
+        /// <param name="segmentData">Segment data.</param>
+        /// <param name="blocksToAdd">Running list of blocks to add.</param>
+        /// <exception cref="IndexOutOfRangeException">Thrown if a depth plane has a bad Z offset.</exception>
+        private void CreateBlocksToAdd(GridPosition segmentIndex, WorldSegmentBlockData segmentData, IList<BlockRecord> blocksToAdd)
+        {
             /* Start by processing depth plaens containing non-default blocks. */
             var processed = new bool[config.SegmentLength];
+            var toAdd = new List<BlockRecord>();
             foreach (var plane in segmentData.DataPlanes)
             {
                 // Validation.
@@ -179,15 +191,15 @@ namespace Sovereign.ClientCore.Systems.WorldManagement
                 }
 
                 processed[plane.OffsetZ] = true;
-                ProcessMixedPlane(plane, segmentData.DefaultMaterialsPerPlane[plane.OffsetZ]);
+                ProcessMixedPlane(segmentIndex, plane, segmentData.DefaultMaterialsPerPlane[plane.OffsetZ], blocksToAdd);
             }
 
             /* Process the remaining all-default planes. */
             for (int i = 0; i < processed.Length; i++)
             {
-                if (!processed[i])
+                if (!processed[i] && segmentData.DefaultMaterialsPerPlane[i].MaterialId != Material.Air)
                 {
-                    ProcessDefaultPlane(i, segmentData.DefaultMaterialsPerPlane[i]);
+                    ProcessDefaultPlane(segmentIndex, i, segmentData.DefaultMaterialsPerPlane[i], blocksToAdd);
                 }
             }
         }
@@ -195,47 +207,99 @@ namespace Sovereign.ClientCore.Systems.WorldManagement
         /// <summary>
         /// Processes a plane containing only default blocks.
         /// </summary>
+        /// <param name="segmentIndex">Segment index.</param>
         /// <param name="offsetZ">Z offset of the plane.</param>
-        /// <param name="defaultBlock">Default block type.</param>"
-        private void ProcessDefaultPlane(int offsetZ, BlockMaterialData defaultBlock)
+        /// <param name="defaultBlock">Default block type.</param>
+        /// <param name="blocksToAdd">List of blocks to add.</param>
+        private void ProcessDefaultPlane(GridPosition segmentIndex, int offsetZ, BlockMaterialData defaultBlock, IList<BlockRecord> blocksToAdd)
         {
-            throw new NotImplementedException();
+            var basePosition = worldSegmentResolver.GetRangeForWorldSegment(segmentIndex).Item1;
+            var baseX = (int)basePosition.X;
+            var baseY = (int)basePosition.Y;
+            var baseZ = (int)basePosition.Z;
+
+            for (int x = 0; x < config.SegmentLength; x++)
+            {
+                for (int y = 0; y < config.SegmentLength; y++)
+                {
+                    var block = new BlockRecord()
+                    {
+                        Position = new GridPosition(baseX + x, baseY + y, baseZ + offsetZ),
+                        Material = defaultBlock.MaterialId,
+                        MaterialModifier = defaultBlock.ModifierId
+                    };
+                    blocksToAdd.Add(block);
+                }
+            }
         }
 
         /// <summary>
         /// Processes a plane containing non-default blocks.
         /// </summary>
+        /// <param name="segmentIndex">Segment index.</param>
         /// <param name="plane">Depth plane to process.</param>
         /// <param name="defaultBlock">Default block type.</param>
-        private void ProcessMixedPlane(WorldSegmentBlockDataPlane plane, BlockMaterialData defaultBlock)
+        /// <param name="blocksToAdd">List of blocks to add.</param>
+        private void ProcessMixedPlane(GridPosition segmentIndex, WorldSegmentBlockDataPlane plane, BlockMaterialData defaultBlock, IList<BlockRecord> blocksToAdd)
         {
-            throw new NotImplementedException();
-        }
+            var basePosition = worldSegmentResolver.GetRangeForWorldSegment(segmentIndex).Item1;
+            var baseX = (int)basePosition.X;
+            var baseY = (int)basePosition.Y;
+            var baseZ = (int)basePosition.Z;
 
-        /// <summary>
-        /// Creates the test blocks for a world segment.
-        /// </summary>
-        /// <param name="blocks">List to hold created blocks.</param>
-        /// <param name="segmentIndex">World segment index.</param>
-        private void CreateBlocks(IList<BlockRecord> blocks,
-            GridPosition segmentIndex)
-        {
-            var origin = (GridPosition)worldSegmentResolver
-                .GetRangeForWorldSegment(segmentIndex).Item1;
-            for (int i = 0; i < config.SegmentLength; ++i)
+            /* Process the non-default blocks. */
+            var nonDefaults = new bool[config.SegmentLength, config.SegmentLength];
+            foreach (var line in plane.Lines)
             {
-                for (int j = 0; j < config.SegmentLength; ++j)
+                /* Validate. */
+                if (line.OffsetY < 0 || line.OffsetY >= config.SegmentLength)
                 {
-                    var pos = new GridPosition(
-                        origin.X + i,
-                        origin.Y + j,
-                        origin.Z);
-                    blocks.Add(new BlockRecord()
+                    throw new IndexOutOfRangeException("Bad Y-offset in world segment data line.");
+                }
+
+                foreach (var block in line.BlockData)
+                {
+                    /* Validate. */
+                    if (block.OffsetX < 0 || block.OffsetX >= config.SegmentLength)
                     {
-                        Position = pos,
-                        Material = 2 + ((i % 2 + j) % 2),
-                        MaterialModifier = 0
-                    });
+                        throw new IndexOutOfRangeException("Bad X-offset in world segment data block.");
+                    }
+
+                    var x = baseX + block.OffsetX;
+                    var y = baseY + line.OffsetY;
+                    var z = baseZ + plane.OffsetZ;
+
+                    /* Add the block. */
+                    var blockRecord = new BlockRecord()
+                    {
+                        Position = new GridPosition(x, y, z),
+                        Material = block.MaterialData.MaterialId,
+                        MaterialModifier = block.MaterialData.ModifierId
+                    };
+                    blocksToAdd.Add(blockRecord);
+
+                    nonDefaults[block.OffsetX, line.OffsetY] = true;
+                }
+            }
+
+            /* Fill in the default blocks (unless they are air). */
+            if (defaultBlock.MaterialId != Material.Air)
+            {
+                for (var i = 0; i < config.SegmentLength; ++i)
+                {
+                    for (var j = 0; j < config.SegmentLength; ++j)
+                    {
+                        if (!nonDefaults[i, j])
+                        {
+                            var blockRecord = new BlockRecord()
+                            {
+                                Position = new GridPosition(baseX + i, baseY + j, baseZ + plane.OffsetZ),
+                                Material = defaultBlock.MaterialId,
+                                MaterialModifier = defaultBlock.ModifierId
+                            };
+                            blocksToAdd.Add(blockRecord);
+                        }
+                    }
                 }
             }
         }